使用新的OpenBLive的sdk进行弹幕获取，削弱荆棘，荆棘反伤由（等级/等级+20）改为（等级/等级+30），爆竹基础攻击力由5调整至3。支持长id玩家加入和正常游戏了

+++ Assets/McPveRoundManager.cs
@@ -11,7 +11,7 @@ public class McPveRoundManager : McRoundManager
 {
     private GameObject fireWall;
     private UnityTimer.Timer fireWallTimer;
-    protected override void ParseGiftInMcMode(SteveCommander steveCommander, string giftName, int battery)
+    protected override void ParseGiftInMcMode(SteveCommander steveCommander, string giftName, long battery)
     {
         base.ParseGiftInMcMode(steveCommander, giftName, battery);
