+++ Assets/Scripts/DebugDialog.cs
@@ -18,7 +18,7 @@ public class DebugDialog : Dialog
     public Text playerName;
 
     public string initName = "云空";
-    public int initUid = 23204263;
+    public long initUid = 23204263;
     
     public int playerIndex = 0;
